refactor: modularize settlement templates
